+++ components/batch-system/src/lib.rs
@@ -4,11 +4,17 @@
 extern crate slog_global;
 #[macro_use]
 extern crate tikv_util;
+#[cfg(feature = "test-runner")]
+#[macro_use]
+extern crate derive_more;
 
 mod batch;
 mod fsm;
 mod mailbox;
 mod router;
+
+#[cfg(feature = "test-runner")]
+pub mod test_runner;
 
 pub use self::batch::{create_system, BatchRouter, BatchSystem, HandlerBuilder, PollHandler};
 pub use self::fsm::Fsm;
